refactor: filestructure scraper
