+++ src/scraper/frameDataPage/tableScrape.mjs
@@ -1,36 +1,38 @@
+import { handleImage, replaceWithAbbreviation } from '../helperFunctions/handleImage.mjs'
+
 function extractDataFromTable() {
 
-  const filenameAbbreviations = {
-    'arrow_3.png': '_',
-    'icon_kick.png': 'K',
-    'icon_kick_h.png': 'HK',
-    'icon_kick_l.png': 'LK',
-    'icon_kick_m.png': 'MK',
-    'icon_punch.png': 'P',
-    'icon_punch_h.png': 'HP',
-    'icon_punch_l.png': 'LP',
-    'icon_punch_m.png': 'MP',
-    'key-all.png': '*',
-    'key-circle.png': 'C',
-    'key-d.png': '2',
-    'key-dl.png': '1',
-    'key-dr.png': '3',
-    'key-l.png': '4',
-    'key-nutral.png': '5',
-    'key-or.png': '|',
-    'key-plus.png': '+',
-    'key-r.png': '6',
-    'key-u.png': '8',
-    'key-ul.png': '7',
-    'key-ur.png': '9',
-    'modern_auto.png': 'AUTO',
-    'modern_dl.png': 'DI',
-    'modern_dp.png': 'DP',
-    'modern_h.png': 'H',
-    'modern_l.png': 'L',
-    'modern_m.png': 'M',
-    'modern_sp.png': 'SP',
-  };
+  // const filenameAbbreviations = {
+  //   'arrow_3.png': '_',
+  //   'icon_kick.png': 'K',
+  //   'icon_kick_h.png': 'HK',
+  //   'icon_kick_l.png': 'LK',
+  //   'icon_kick_m.png': 'MK',
+  //   'icon_punch.png': 'P',
+  //   'icon_punch_h.png': 'HP',
+  //   'icon_punch_l.png': 'LP',
+  //   'icon_punch_m.png': 'MP',
+  //   'key-all.png': '*',
+  //   'key-circle.png': 'C',
+  //   'key-d.png': '2',
+  //   'key-dl.png': '1',
+  //   'key-dr.png': '3',
+  //   'key-l.png': '4',
+  //   'key-nutral.png': '5',
+  //   'key-or.png': '|',
+  //   'key-plus.png': '+',
+  //   'key-r.png': '6',
+  //   'key-u.png': '8',
+  //   'key-ul.png': '7',
+  //   'key-ur.png': '9',
+  //   'modern_auto.png': 'AUTO',
+  //   'modern_dl.png': 'DI',
+  //   'modern_dp.png': 'DP',
+  //   'modern_h.png': 'H',
+  //   'modern_l.png': 'L',
+  //   'modern_m.png': 'M',
+  //   'modern_sp.png': 'SP',
+  // };
 
   const moveTypes = ['Normal Moves', 'Unique Attacks', 'Special Moves', 'Super Arts', 'Throws', 'Common Moves'];
   
@@ -68,26 +70,26 @@ function extractDataFromTable() {
     return images;
   };
 
-  const extractFileName = (url) => {
-    const regexSelectImgFileName = /([^\/]+)$/;
-    const match = url.match(regexSelectImgFileName);
-    return match ? match[0] : null;
-  };
+  // const extractFileName = (url) => {
+  //   const regexSelectImgFileName = /([^\/]+)$/;
+  //   const match = url.match(regexSelectImgFileName);
+  //   return match ? match[0] : null;
+  // };
   
-  const handleImage = (element) => {
-    if (element.tagName.toLowerCase() === 'img' && element.src) {
-      const filename = extractFileName(element.src);
-      return filename ? [filename] : [];
-    }
-    return [];
-  };
+  // const handleImage = (element) => {
+  //   if (element.tagName.toLowerCase() === 'img' && element.src) {
+  //     const filename = extractFileName(element.src);
+  //     return filename ? [filename] : [];
+  //   }
+  //   return [];
+  // };
   
-  const replaceWithAbbreviation = (arrayOfFilenames) => {
-    return arrayOfFilenames.map(filename => {
-      if (!filename) return filename;
-      return filenameAbbreviations[filename] || filename;
-    });
-  };
+  // const replaceWithAbbreviation = (arrayOfFilenames) => {
+  //   return arrayOfFilenames.map(filename => {
+  //     if (!filename) return filename;
+  //     return filenameAbbreviations[filename] || filename;
+  //   });
+  // };
   
   rows.forEach(row => {
