+++ src/scraper/helperFunctions/handleImage.mjs
